Refactor Position::from() for brevity
Reordering of the usage of the fen tokens precludes the need for
creating 'storage' variables. Switched to retrieving the fen token and
processing it fully before fetching the next token. Also switched
variable names to be the names of the related Position field to take
advantage of the cleaner field init shorthand.