Improve sq_to_bitboard performance
Using to_digit() causes unnecessary function call and logic overhead
when the bounds of both file and rank are known ahead of time. Casting
to u32 and subracting is how the standard library implements this
functionality but with extra logic around the radix and returning an
Option<T> instead of the value we really wanted.